:art: refactor : BlackJackGame에서 플레이어들을 일급 콜렉션으로 포장
